requirements.txt
blinker==1.9.0
click==8.1.8
cmdstanpy==1.2.5
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
Deprecated==1.2.18
dotenv==0.9.9
Flask==3.1.0
flask-cors==5.0.1
Flask-Limiter==3.11.0
fonttools==4.56.0
holidays==0.71
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
limits==4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
numpy==2.2.5
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
prophet==1.1.6
psycopg2-binary==2.9.10
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
stanio==0.5.1
statsmodels==0.14.4
tabulate==0.9.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
